feat: add login component
The code changes add a new `Login` component to the client application. This component allows users to enter their email and password and click a login button. The component uses React's `useState` hook to manage the state of the input fields. When the login button is clicked, a `login` function is called, which currently logs a message to the console.

The suggested commit message is "feat: add login component".